move generic analytics functions into a module
