+++ modules/analytics/index.js
@@ -1,0 +1,2 @@
+// @flow
+export * from './google-analytics'
